pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperkb"
dynamic = ["version"]
description = "A hyperconnected knowledge base with hybrid search"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "httpx>=0.24",
]

[project.optional-dependencies]
crypto = ["cryptography>=42.0"]
mcp = ["mcp[cli]>=1.0"]
sync = ["boto3>=1.34", "watchdog>=4.0"]
all = [
    "cryptography>=42.0",
    "mcp[cli]>=1.0",
    "boto3>=1.34",
    "watchdog>=4.0",
]
dev = ["pytest>=7.0", "moto[s3]>=5.0"]

[project.scripts]
hkb = "hyperkb.cli:cli"
hkb-mcp = "hyperkb.mcp_server:main"

[tool.setuptools.packages.find]
include = ["hyperkb*"]

[tool.setuptools_scm]
version_file = "hyperkb/_version.py"
